Struct sharing_is_caring::server::ClientHandler

pub struct ClientHandler<U, A, R> { /* fields omitted */ }

A struct that contains the implementation for behavior associated with a single client, like when responding to a single client's requests. For now, all client handlers are created equal because we treat all our clients equally, and so the fields in these objects are all the same. However, this struct gives room for expansion.

See Server::as_client_handler.