Struct sharing_is_caring::server::Server

pub struct Server<U, A, R> { /* fields omitted */ }

The server object that defines behavior during the lifetime of our program.

Methods

impl<U: Universe<A>, A: Atom, R: Responder<U>> Server<U, A, R>

pub fn new(universe: U, responder: R) -> Self

Creates a new server with the given universe, name of our server, and a function to generate a response whenever we receive a request.

pub fn as_client_handler(&self, out: Sender) -> ClientHandler<U, A, R>

Creates a ClientHandler instance from the object.

pub fn listen(self, host: &str) -> Result<()>